+++ src/main/java/soma/edupi/db/EdupiDbApplication.java
@@ -1,0 +1,13 @@
+package soma.edupi.db;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class EdupiDbApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(EdupiDbApplication.class, args);
+    }
+
+}
