+++ src/main/java/soma/edupi/db/classroom/domain/Classroom.java
@@ -1,0 +1,48 @@
+package soma.edupi.db.classroom.domain;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.validation.constraints.NotNull;
+import java.time.LocalDateTime;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+@Entity
+@Getter
+@NoArgsConstructor
+public class Classroom {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "classroom_id", updatable = false)
+    private Long id;
+
+    @NotNull
+    private Long userId;
+
+    @NotNull
+    private String name;
+
+    @NotNull
+    private String inviteLink;
+
+    @CreationTimestamp
+    @Column(updatable = false)
+    private LocalDateTime createdAt;
+
+    @UpdateTimestamp
+    private LocalDateTime updatedAt;
+
+    @Builder
+    public Classroom(Long userId, String name, String inviteLink) {
+        this.userId = userId;
+        this.name = name;
+        this.inviteLink = inviteLink;
+    }
+}
