+++ src/main/java/soma/edupi/db/common/models/response/DefaultErrorResponse.java
@@ -1,0 +1,11 @@
+package soma.edupi.db.common.models.response;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class DefaultErrorResponse {
+
+    private final String message;
+}
