+++ src/main/java/soma/edupi/db/member/exception/ServerException.java
@@ -1,0 +1,11 @@
+package soma.edupi.db.member.exception;
+
+/**
+ * 서버에서 처리 실패에 대한 Exception
+ */
+public class ServerException extends RuntimeException {
+
+    public ServerException(String message) {
+        super(message);
+    }
+}
