+++ src/main/java/soma/edupi/db/member/models/request/LoginRequest.java
@@ -1,0 +1,23 @@
+package soma.edupi.db.member.models.request;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
+import jakarta.validation.constraints.Size;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class LoginRequest {
+
+    @NotBlank(message = "이메일은 공백이 아니어야 합니다.")
+    @Email(message = "잘못된 이메일 형식입니다.")
+    private String email;
+
+    @NotBlank(message = "비밀번호는 공백이 아니어야 합니다.")
+    @Size(min = 8, max = 20, message = "비밀번호는 8에서 20자 이내여야 합니다.")
+    @Pattern(regexp = "^(?=.*[!@#$%^&*()_+\\-=\\[\\]{};':\"\\\\|,.<>/?]).*$", message = "비밀번호는 최소 1개의 특수기호를 포함해야 합니다.")
+    private String password;
+
+}
