--- src/main/java/soma/haeya/edupi_db/EdupiDbApplication.java
@@ -1,13 +1,0 @@
-package soma.haeya.edupi_db;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EdupiDbApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(EdupiDbApplication.class, args);
-    }
-
-}
