--- src/main/java/soma/haeya/edupi_db/common/models/response/DefaultErrorResponse.java
@@ -1,11 +1,0 @@
-package soma.haeya.edupi_db.common.models.response;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class DefaultErrorResponse {
-
-    private final String message;
-}
